Add hook components
A component is part of a hook. It can be a request validator, filter or
action. Each component has their own storage they can share between
instances of that component. Every component added to a hook has a
unique identifier.

Whenever a request comes in, it should be processed by each individual
component in order. If one of the components returns an error during
processing, the request is dropped.

The Github validator component makes sure the signature in the request
header matches the HMAC of the body with your configured secret. The log
action component logs the incoming request to stderr. The write file
component writes the request contents to a file on disk.

Refs #1, refs #4